Adapted to CLI
