+++ src/scripts/generate_datasets.py
@@ -2,6 +2,7 @@ from src.data.data_extract import DataExtractor
 from src.data.data_preprocess import DataPreprocessor
 from src.data.constants import Currency
 from typing import NoReturn
+from src.scripts.click_utils import CurrencyType
 
 import logging
 import click
@@ -11,21 +12,13 @@ logging.basicConfig(
     level=logging.INFO,  
     format='%(asctime)s | %(levelname)s | %(message)s')
 
-
-class CurrencyType(click.ParamType):
-    def convert(self, value, param, ctx):
-        if isinstance(value, str):
-            try:
-                return Currency(value).name
-            except ValueError:
-                self.fail(f"{value!r} is not a valid string", param, ctx)
-
 CURRENCY_TYPE = CurrencyType()
+
 
 @click.command()
 @click.argument('base', type=CURRENCY_TYPE)
 @click.argument('quote', type=CURRENCY_TYPE)
-@click.option('--clobber', defautl=False, type=click.BOOL,
+@click.option('--clobber', default=False, type=click.BOOL,
               help="If data should be re-generated or not")
 def process_fx_pair(
     base: Currency, 
@@ -52,7 +45,6 @@ def process_fx_pair(
 
 
 if __name__ == '__main__':
-    process_fx_pair(Currency.AUD, Currency.JPY)
     process_fx_pair(Currency.AUD, Currency.JPY)
     process_fx_pair(Currency.AUD, Currency.USD)
     process_fx_pair(Currency.CAD, Currency.JPY)
